New Makefile, LAPACK95 removed, misc updates

+++ collisions/collisions.f90
@@ -4,7 +4,7 @@ use phys_consts
 implicit none
 
 !THESE DATA BLOCKS ARE UGLY; COULD THERE BE A BETTER WAY?
-real(8), dimension(lsp,ln) :: Csn
+real(wp), dimension(lsp,ln) :: Csn
 data Csn(1,:) /-1.0, 6.82d-10, 6.64d-10, -1.0/
 data Csn(2,:) /2.44d-10, 4.34d-10, 4.27d-10, 0.69d-10/
 data Csn(3,:) /2.58d-10, -1.0, 4.49d-10, 0.74d-10/
@@ -13,7 +13,7 @@ data Csn(5,:) /4.42d-10, 7.47d-10, 7.25d-10, 1.45d-10/
 data Csn(6,:) /-1.0, 33.6d-10, 32.0d-10, -1.0/
 data Csn(7,:) /-1.0, -1.0, -1.0, -1.0/
 
-real(8), dimension(lsp,ln) :: C2sn1
+real(wp), dimension(lsp,ln) :: C2sn1
 data C2sn1(1,:) /3.67d-11, 0.0, 0.0, 4.63d-12/
 data C2sn1(2,:) /0.0, 0.0, 0.0, 0.0/
 data C2sn1(3,:) /0.0, 5.14d-11, 0.0, 0.0/
@@ -22,7 +22,7 @@ data C2sn1(5,:) /0.0, 0.0, 0.0, 0.0/
 data C2sn1(6,:) /6.61d-11, 0.0, 0.0, 2.65d-10/
 data C2sn1(7,:) /-1.0, -1.0, -1.0, -1.0/
 
-real(8), dimension(lsp,ln) :: C2sn2
+real(wp), dimension(lsp,ln) :: C2sn2
 data C2sn2(1,:) /0.064, 0.0, 0.0, -1.0/
 data C2sn2(2,:) /0.0, 0.0, 0.0, 0.0/
 data C2sn2(3,:) /0.0, 0.069, 0.0, 0.0/
@@ -31,7 +31,7 @@ data C2sn2(5,:) /0.0, 0.0, 0.0, 0.0/
 data C2sn2(6,:) /0.047, 0.0, 0.0, 0.083/
 data C2sn2(7,:) /-1.0, -1.0, -1.0, -1.0/
 
-real(8), dimension(lsp,lsp) :: Csj
+real(wp), dimension(lsp,lsp) :: Csj
 data Csj(1,:) /0.22, 0.26, 0.25, 0.26, 0.22, 0.077, 1.87e-3/
 data Csj(2,:) /0.14, 0.16, 0.16, 0.17, 0.13, 0.042, 9.97e-4/
 data Csj(3,:) /0.15, 0.17, 0.17, 0.18, 0.14, 0.045, 1.07e-3/
@@ -52,16 +52,16 @@ contains
     !-------Note that it is done on a per species basis
 
     integer, intent(in) :: isp,isp2
-    real(8), dimension(:,:,:,:), intent(in) :: nn
-    real(8), dimension(:,:,:), intent(in) :: Tn
-    real(8), dimension(-1:,-1:,-1:,:), intent(in) :: Ts
+    real(wp), dimension(:,:,:,:), intent(in) :: nn
+    real(wp), dimension(:,:,:), intent(in) :: Tn
+    real(wp), dimension(-1:,-1:,-1:,:), intent(in) :: Ts
 
-    real(8), dimension(1:size(Tn,1),1:size(Tn,2),1:size(Tn,3)), &
+    real(wp), dimension(1:size(Tn,1),1:size(Tn,2),1:size(Tn,3)), &
               intent(out) :: nusn
 
     integer :: lx1,lx2,lx3
-    real(8) :: mred
-    real(8),dimension(1:size(Tn,1),1:size(Tn,2),1:size(Tn,3)) :: Teff
+    real(wp) :: mred
+    real(wp),dimension(1:size(Tn,1),1:size(Tn,2),1:size(Tn,3)) :: Teff
 
     lx1=size(Ts,1)-4
     lx2=size(Ts,2)-4
@@ -110,14 +110,14 @@ contains
     !-------Note that it is done on a per species basis
 
     integer, intent(in) :: isp,isp2
-    real(8), dimension(-1:,-1:,-1:,:), intent(in) :: ns,Ts,vs1
+    real(wp), dimension(-1:,-1:,-1:,:), intent(in) :: ns,Ts,vs1
 
-    real(8), dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4), &
+    real(wp), dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4), &
               intent(out) :: nusj,Phisj,Psisj
 
     integer :: lx1,lx2,lx3
-    real(8) :: mred
-    real(8),dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4) &
+    real(wp) :: mred
+    real(wp),dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4) &
               :: Teff,Wsj,Phitmp
 
 
@@ -156,12 +156,12 @@ contains
     !-------Note that it is done on a per species basis
 
     integer, intent(in) :: isp
-    real(8), dimension(-1:,-1:,-1:), intent(in) :: Ts,ns
+    real(wp), dimension(-1:,-1:,-1:), intent(in) :: Ts,ns
 
-    real(8), dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4,ln), intent(in) :: nn
-    real(8), dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4), intent(in) :: J1
+    real(wp), dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4,ln), intent(in) :: nn
+    real(wp), dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4), intent(in) :: J1
 
-    real(8), dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4), intent(out) :: lambda,beta
+    real(wp), dimension(1:size(Ts,1)-4,1:size(Ts,2)-4,1:size(Ts,3)-4), intent(out) :: lambda,beta
 
     integer :: lx1,lx2,lx3
 
@@ -193,16 +193,16 @@ contains
     !-------VARS. INCLUDE GHOST CELLS
     !------------------------------------------------------------
 
-    real(8), dimension(:,:,:,:), intent(in) :: nn
-    real(8), dimension(:,:,:), intent(in) :: Tn
-    real(8), dimension(-1:,-1:,-1:,:), intent(in) :: ns,Ts,vs1
-    real(8), dimension(-1:,-1:,-1:), intent(in) :: B1
-    real(8), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4), intent(out) :: sig0,sigP,sigH
-    real(8), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4,lsp), intent(out) :: muP,muH,muPvn,muHvn
+    real(wp), dimension(:,:,:,:), intent(in) :: nn
+    real(wp), dimension(:,:,:), intent(in) :: Tn
+    real(wp), dimension(-1:,-1:,-1:,:), intent(in) :: ns,Ts,vs1
+    real(wp), dimension(-1:,-1:,-1:), intent(in) :: B1
+    real(wp), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4), intent(out) :: sig0,sigP,sigH
+    real(wp), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4,lsp), intent(out) :: muP,muH,muPvn,muHvn
 
     integer :: isp,isp2,lx1,lx2,lx3
-    real(8), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4) :: OMs
-    real(8), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4) :: nu,nuej,Phisj,Psisj,nutmp,mupar,mubase,rho
+    real(wp), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4) :: OMs
+    real(wp), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4) :: nu,nuej,Phisj,Psisj,nutmp,mupar,mubase,rho
 
 
     lx1=size(Ts,1)-4
@@ -266,11 +266,11 @@ contains
     !-------DENSITY/MAG FIELD STATE VARIABLE INCLUDES GHOST CELLS.
     !------------------------------------------------------------
 
-    real(8), dimension(-1:,-1:,-1:,:), intent(in) :: ns
-    real(8), dimension(-1:,-1:,-1:), intent(in) :: B1
+    real(wp), dimension(-1:,-1:,-1:,:), intent(in) :: ns
+    real(wp), dimension(-1:,-1:,-1:), intent(in) :: B1
     integer, intent(in) :: flagcap
 
-    real(8), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4), intent(out) :: incap
+    real(wp), dimension(1:size(ns,1)-4,1:size(ns,2)-4,1:size(ns,3)-4), intent(out) :: incap
 
     integer :: lx1,lx2,lx3,isp
